Make text strokeWidth a double.

+++ LottieUWP/Model/DocumentData.cs
@@ -14,10 +14,10 @@ namespace LottieUWP.Model
         internal readonly double BaselineShift;
         internal readonly Color Color;
         internal readonly Color StrokeColor;
-        internal readonly int StrokeWidth;
+        internal readonly double StrokeWidth;
         internal readonly bool StrokeOverFill;
 
-        internal DocumentData(string text, string fontName, double size, int justification, int tracking, double lineHeight, double baselineShift, Color color, Color strokeColor, int strokeWidth, bool strokeOverFill)
+        internal DocumentData(string text, string fontName, double size, int justification, int tracking, double lineHeight, double baselineShift, Color color, Color strokeColor, double strokeWidth, bool strokeOverFill)
         {
             Text = text;
             FontName = fontName;
